chore: reorganize configs
