+++ resources/software.json
@@ -26,7 +26,7 @@
         "repositories": [
             {
                 "type": "github",
-                "uri": "temporalio/sdk-php",
+                "uri": "temporalio/cli",
                 "asset-pattern": "/^temporal-.*/"
             }
         ],
