+++ src/DLoad.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Internal\DLoad;
 
 use Internal\DLoad\Module\Archive\ArchiveFactory;
-use Internal\DLoad\Module\Common\Config\Destination;
 use Internal\DLoad\Module\Common\Config\Embed\File;
 use Internal\DLoad\Module\Common\Config\Embed\Software;
+use Internal\DLoad\Module\Common\Input\Destination;
 use Internal\DLoad\Module\Downloader\Downloader;
 use Internal\DLoad\Module\Downloader\SoftwareCollection;
 use Internal\DLoad\Module\Downloader\Task\DownloadResult;
