+++ src/Module/Common/Architecture.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Internal\DLoad\Module\Common;
 
-use Internal\DLoad\Module\Common\Config\BuildInput;
+use Internal\DLoad\Module\Common\Input\Build;
 use Internal\DLoad\Service\Factoriable;
 
 /**
@@ -19,7 +19,7 @@ enum Architecture: string implements Factoriable
 
     private const ERROR_UNKNOWN_ARCH = 'Current architecture `%s` may not be supported.';
 
-    public static function create(BuildInput $config): self
+    public static function create(Build $config): self
     {
         return self::tryFrom((string) $config->arch) ?? self::fromGlobals();
     }
