+++ src/Module/Common/Config/Downloader.php
@@ -4,7 +4,10 @@ declare(strict_types=1);
 
 namespace Internal\DLoad\Module\Common\Config;
 
-final class DownloaderConfig
+use Internal\DLoad\Module\Common\Internal\Attribute\XPath;
+
+final class Downloader
 {
+    #[XPath('/dload/@temp-dir')]
     public ?string $tmpDir = null;
 }
