+++ src/Module/Common/Config/Embed/Repository.php
@@ -6,6 +6,13 @@ namespace Internal\DLoad\Module\Common\Config\Embed;
 
 use Internal\DLoad\Module\Common\Internal\Attribute\XPath;
 
+/**
+ * @psalm-type RepositoryArray = array{
+ *     type: non-empty-string,
+ *     uri: non-empty-string,
+ *     asset-pattern?: non-empty-string
+ * }
+ */
 final class Repository
 {
     #[XPath('@type')]
@@ -17,6 +24,9 @@ final class Repository
     #[XPath('@asset-pattern')]
     public string $assetPattern = '/^.*$/';
 
+    /**
+     * @param RepositoryArray $repositoryArray
+     */
     public static function fromArray(mixed $repositoryArray): self
     {
         $self = new self();
