+++ src/Module/Common/Input/Destination.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Internal\DLoad\Module\Common\Config;
+namespace Internal\DLoad\Module\Common\Input;
 
 use Internal\DLoad\Module\Common\Internal\Attribute\InputOption;
