+++ src/Module/Common/OperatingSystem.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Internal\DLoad\Module\Common;
 
-use Internal\DLoad\Module\Common\Config\BuildInput;
+use Internal\DLoad\Module\Common\Input\Build;
 use Internal\DLoad\Service\Factoriable;
 
 /**
@@ -22,7 +22,7 @@ enum OperatingSystem: string implements Factoriable
 
     private const ERROR_UNKNOWN_OS = 'Current OS `%s` may not be supported';
 
-    public static function create(BuildInput $config): static
+    public static function create(Build $config): static
     {
         return self::tryFrom((string) $config->os) ?? self::fromGlobals();
     }
