+++ src/Module/Repository/Internal/GitHub/Factory.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Internal\DLoad\Module\Repository\Internal\GitHub;
 
-use Internal\DLoad\Module\Common\Config\GitHubConfig;
+use Internal\DLoad\Module\Common\Config\GitHub as GitHubConfig;
 use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
@@ -15,7 +15,7 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 final class Factory
 {
     public function __construct(
-        private readonly GitHubConfig $config,
+        private readonly GitHub $config,
     ) {}
 
     /**
